Add case submitter
