+++ cg/meta/orders/api.py
@@ -17,11 +17,14 @@ from cg.constants import DataDelivery, Pipeline
 from cg.exc import OrderError
 from cg.models.orders.order import OrderIn, OrderType
 from cg.store import Store, models
+from .balsamic_submitter import BalsamicSubmitter
 from .fastq_submitter import FastqSubmitter
 
 from .lims import process_lims
 from .metagenome_submitter import MetagenomeSubmitter
 from .microsalt_submitter import MicrosaltSubmitter
+from .mip_dna_submitter import MipDnaSubmitter
+from .mip_rna_submitter import MipRnaSubmitter
 from .sars_cov_2_submitter import SarsCov2Submitter
 from .status import StatusHandler
 from .submitter import Submitter
@@ -39,6 +42,9 @@ def _get_submit_handler(project: OrderType, lims: LimsAPI, status: Store) -> Sub
         OrderType.METAGENOME: MetagenomeSubmitter,
         OrderType.MICROSALT: MicrosaltSubmitter,
         OrderType.SARS_COV_2: SarsCov2Submitter,
+        OrderType.MIP_DNA: MipDnaSubmitter,
+        OrderType.BALSAMIC: BalsamicSubmitter,
+        OrderType.MIP_RNA: MipRnaSubmitter,
     }
     if project in submitters:
         return submitters[project](lims=lims, status=status)
@@ -59,11 +65,9 @@ class OrdersAPI(StatusHandler):
 
         Main entry point for the class towards interfaces that implements it.
         """
-        submit_handler = _get_submit_handler(project, lims=self.lims, status=self.status)
+        submit_handler: Submitter = _get_submit_handler(project, lims=self.lims, status=self.status)
         if submit_handler:
             submit_handler.validate_order(order=order_in)
-        else:
-            self.validate_order(order_in, project)
 
         # detect manual ticket assignment
         ticket_number: Optional[int] = TicketHandler.parse_ticket_number(order_in.name)
@@ -79,15 +83,6 @@ class OrdersAPI(StatusHandler):
 
         order_func = self._get_submit_func(project.value)
         return order_func(order_in)
-
-    def validate_order(self, order_in: OrderIn, project: OrderType):
-        self._validate_samples_available_to_customer(
-            project=project, samples=order_in.samples, customer_id=order_in.customer
-        )
-        self._validate_case_names_are_unique(
-            project=project, samples=order_in.samples, customer_id=order_in.customer
-        )
-        self._validate_subject_sex(samples=order_in.samples, customer_id=order_in.customer)
 
     def _submit_fluffy(self, order: OrderIn) -> dict:
         """Submit a batch of ready made libraries for FLUFFY analysis."""
@@ -113,19 +108,6 @@ class OrdersAPI(StatusHandler):
         )
         return {"project": project_data, "records": new_records}
 
-    def _submit_case_samples(self, order: OrderIn, project: OrderType) -> dict:
-        """Submit a batch of samples for sequencing and analysis."""
-        result = self._process_case_samples(order=order, project=project)
-        for case_obj in result["records"]:
-            LOG.info(f"{case_obj.name}: submit family samples")
-            status_samples = [
-                link_obj.sample
-                for link_obj in case_obj.links
-                if link_obj.sample.ticket_number == order.ticket
-            ]
-            self._add_missing_reads(status_samples)
-        return result
-
     def _submit_mip_dna(self, order: OrderIn) -> dict:
         """Submit a batch of samples for sequencing and analysis."""
         return self._submit_case_samples(order=order, project=OrderType.MIP_DNA)
@@ -137,31 +119,6 @@ class OrdersAPI(StatusHandler):
     def _submit_mip_rna(self, order: OrderIn) -> dict:
         """Submit a batch of samples for sequencing and analysis."""
         return self._submit_case_samples(order=order, project=OrderType.MIP_RNA)
-
-    def _process_case_samples(self, order: OrderIn, project: OrderType) -> dict:
-        """Process samples to be analyzed."""
-        project_data = lims_map = None
-
-        # submit new samples to lims
-        new_samples = [sample for sample in order.samples if sample.internal_id is None]
-        if new_samples:
-            project_data, lims_map = process_lims(
-                lims_api=self.lims, lims_order=order, new_samples=new_samples
-            )
-
-        status_data = self.cases_to_status(order=order, project=project)
-        samples = [sample for family in status_data["families"] for sample in family["samples"]]
-        if lims_map:
-            self._fill_in_sample_ids(samples, lims_map)
-
-        new_families = self.store_cases(
-            customer=status_data["customer"],
-            order=status_data["order"],
-            ordered=project_data["date"] if project_data else dt.datetime.now(),
-            ticket=order.ticket,
-            cases=status_data["families"],
-        )
-        return {"project": project_data, "records": new_families}
 
     def _add_missing_reads(self, samples: List[models.Sample]):
         """Add expected reads/reads missing."""
@@ -180,53 +137,6 @@ class OrdersAPI(StatusHandler):
                 LOG.info(f"{sample['name']} -> {internal_id}: connect sample to LIMS")
                 sample[id_key] = internal_id
 
-    def _validate_samples_available_to_customer(
-        self, project: OrderType, samples: List[OrderInSample], customer_id: str
-    ) -> None:
-        """Validate that the customer have access to all samples"""
-        if project not in (
-            OrderType.BALSAMIC,
-            OrderType.MIP_DNA,
-            OrderType.MIP_RNA,
-        ):
-            return
-
-        sample: Of1508Sample
-        for sample in samples:
-            if not sample.internal_id:
-                continue
-
-            existing_sample: models.Sample = self.status.sample(sample.internal_id)
-            data_customer: models.Customer = self.status.customer(customer_id)
-
-            if existing_sample.customer.customer_group_id != data_customer.customer_group_id:
-                raise OrderError(f"Sample not available: {sample.name}")
-
-    def _validate_case_names_are_unique(
-        self, project: OrderType, samples: List[OrderInSample], customer_id: str
-    ) -> None:
-        """Validate that the names of all cases are unused for all samples"""
-        customer_obj: models.Customer = self.status.customer(customer_id)
-        if project not in (
-            OrderType.BALSAMIC,
-            OrderType.MIP_DNA,
-            OrderType.MIP_RNA,
-        ):
-            return
-
-        sample: Of1508Sample
-        for sample in samples:
-
-            if self._rerun_of_existing_case(sample):
-                continue
-
-            if self.status.find_family(customer=customer_obj, name=sample.family_name):
-                raise OrderError(f"Case name {sample.family_name} already in use")
-
-    @staticmethod
-    def _rerun_of_existing_case(sample: Of1508Sample) -> bool:
-        return sample.case_internal_id is not None
-
     def _get_submit_func(self, project_type: OrderType) -> typing.Callable:
         """Get the submit method to call for the given type of project"""
 
@@ -238,38 +148,3 @@ class OrdersAPI(StatusHandler):
             return getattr(self, "_submit_sars_cov_2")
 
         return getattr(self, f"_submit_{str(project_type)}")
-
-    def _validate_subject_sex(self, samples: [Of1508Sample], customer_id: str):
-        """Validate that sex is consistent with existing samples, skips samples of unknown sex
-
-        Args:
-            samples     (list[dict]):   Samples to validate
-            customer_id (str):          Customer that the samples belong to
-        Returns:
-            Nothing
-        """
-
-        sample: Of1508Sample
-        for sample in samples:
-            if not isinstance(sample, Of1508Sample):
-                continue
-
-            subject_id: str = sample.subject_id
-            if not subject_id:
-                continue
-            new_gender: str = sample.sex
-            if new_gender == "unknown":
-                continue
-            existing_samples: [models.Sample] = self.status.samples_by_subject_id(
-                customer_id=customer_id, subject_id=subject_id
-            )
-            existing_sample: models.Sample
-            for existing_sample in existing_samples:
-                previous_gender = existing_sample.sex
-                if previous_gender == "unknown":
-                    continue
-
-                if previous_gender != new_gender:
-                    raise OrderError(
-                        f"Sample gender inconsistency for subject_id: {subject_id}: previous gender {previous_gender}, new gender {new_gender}"
-                    )
