+++ cg/meta/orders/balsamic_submitter.py
@@ -1,0 +1,5 @@
+from cg.meta.orders.case_submitter import CaseSubmitter
+
+
+class BalsamicSubmitter(CaseSubmitter):
+    pass
