+++ cg/meta/orders/microbial_submitter.py
@@ -1,3 +1,4 @@
+import datetime as dt
 from typing import List
 
 from cgmodels.cg.constants import Pipeline
@@ -7,10 +8,8 @@ from cg.exc import OrderError
 from cg.meta.orders.lims import process_lims
 from cg.meta.orders.submitter import Submitter
 from cg.models.orders.order import OrderIn
-from cg.models.orders.sample_base import StatusEnum
 from cg.models.orders.samples import MicrobialSample
 from cg.store import models
-import datetime as dt
 
 
 class MicrobialSubmitter(Submitter):
