+++ cg/meta/orders/microsalt_submitter.py
@@ -1,16 +1,4 @@
-from typing import List
-
-from cgmodels.cg.constants import Pipeline
-
-from cg.constants import DataDelivery
-from cg.exc import OrderError
-from cg.meta.orders.lims import process_lims
 from cg.meta.orders.microbial_submitter import MicrobialSubmitter
-from cg.meta.orders.submitter import Submitter
-from cg.models.orders.order import OrderIn
-from cg.models.orders.sample_base import StatusEnum
-from cg.store import models
-import datetime as dt
 
 
 class MicrosaltSubmitter(MicrobialSubmitter):
