+++ tests/meta/orders/test_meta_orders_api.py
@@ -4,6 +4,7 @@ import pytest
 from cg.constants import DataDelivery
 from cg.exc import OrderError
 from cg.meta.orders import OrdersAPI
+from cg.meta.orders.mip_dna_submitter import MipDnaSubmitter
 from cg.models.orders.order import OrderIn, OrderType
 from cg.models.orders.samples import MipDnaSample
 from cg.store import models, Store
@@ -13,7 +14,14 @@ from tests.store_helpers import StoreHelpers
 
 PROCESS_LIMS_FUNCTION_OLD = "cg.meta.orders.api.process_lims"
 PROCESS_LIMS_FUNCTION = "cg.meta.orders.lims.process_lims"
-SUBMITTERS = ["api", "lims", "fastq_submitter", "metagenome_submitter", "microbial_submitter"]
+SUBMITTERS = [
+    "api",
+    "lims",
+    "fastq_submitter",
+    "metagenome_submitter",
+    "microbial_submitter",
+    "case_submitter",
+]
 
 
 def test_too_long_order_name():
@@ -289,12 +297,12 @@ def test_validate_sex_inconsistent_sex(
         store.add_commit(sample_obj)
         assert sample_obj.sex != sample.sex
 
+    submitter: MipDnaSubmitter = MipDnaSubmitter(lims=orders_api.lims, status=orders_api.status)
+
     # WHEN calling _validate_sex
     # THEN an OrderError should be raised on non-matching sex
     with pytest.raises(OrderError):
-        orders_api._validate_subject_sex(
-            samples=order_data.samples, customer_id=order_data.customer
-        )
+        submitter._validate_subject_sex(samples=order_data.samples, customer_id=order_data.customer)
 
 
 def test_validate_sex_consistent_sex(
@@ -318,8 +326,10 @@ def test_validate_sex_consistent_sex(
         store.add_commit(sample_obj)
         assert sample_obj.sex == sample.sex
 
+    submitter: MipDnaSubmitter = MipDnaSubmitter(lims=orders_api.lims, status=orders_api.status)
+
     # WHEN calling _validate_sex
-    orders_api._validate_subject_sex(samples=order_data.samples, customer_id=order_data.customer)
+    submitter._validate_subject_sex(samples=order_data.samples, customer_id=order_data.customer)
 
     # THEN no OrderError should be raised on non-matching sex
 
@@ -346,8 +356,10 @@ def test_validate_sex_unknown_existing_sex(
         store.add_commit(sample_obj)
         assert sample_obj.sex != sample.sex
 
+    submitter: MipDnaSubmitter = MipDnaSubmitter(lims=orders_api.lims, status=orders_api.status)
+
     # WHEN calling _validate_sex
-    orders_api._validate_subject_sex(samples=order_data.samples, customer_id=order_data.customer)
+    submitter._validate_subject_sex(samples=order_data.samples, customer_id=order_data.customer)
 
     # THEN no OrderError should be raised on non-matching sex
 
@@ -376,8 +388,10 @@ def test_validate_sex_unknown_new_sex(
     for sample in order_data.samples:
         assert sample_obj.sex != sample.sex
 
+    submitter: MipDnaSubmitter = MipDnaSubmitter(lims=orders_api.lims, status=orders_api.status)
+
     # WHEN calling _validate_sex
-    orders_api._validate_subject_sex(samples=order_data.samples, customer_id=order_data.customer)
+    submitter._validate_subject_sex(samples=order_data.samples, customer_id=order_data.customer)
 
     # THEN no OrderError should be raised on non-matching sex
